# documentation_evidence/sahmi-current-evidence-2026-08-14/openapi-schema.yml
openapi: 3.0.3
info:
  title: Sahmi API
  version: 1.0.0
  description: Django REST API for the Sahmi investment platform.
paths:
...
  /api/v1/admin/projects/{id}/set-status/:
    post:
      operationId: api_v1_admin_projects_set_status_create
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        description: A UUID string identifying this project.
        required: true
      tags:
      - api
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminProject'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/AdminProject'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/AdminProject'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminProject'
          description: ''
...
  /api/v1/projects/{slug}/set-status/:
    post:
      operationId: api_v1_projects_set_status_create
      parameters:
      - in: path
        name: slug
        schema:
          type: string
        required: true
      tags:
      - api
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Project'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Project'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Project'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
          description: ''
...
components:
  schemas:
...
    AdminProject:
      type: object
      description: |-
        Full serializer used by the owner (``entrepreneur``) and staff for create /
        retrieve / update / destroy of their own project. ``entrepreneur`` is read-only.
      properties:
        clear_business_plan:
          type: boolean
          writeOnly: true
          default: false
        clear_financial_projections:
          type: boolean
          writeOnly: true
          default: false
        clear_ownership_proof:
          type: boolean
          writeOnly: true
          default: false
        clear_cover_image:
          type: boolean
          writeOnly: true
          default: false
        id:
          type: string
          format: uuid
          readOnly: true
        entrepreneur:
          type: string
          format: uuid
        entrepreneur_detail:
          allOf:
          - $ref: '#/components/schemas/AdminRelatedUser'
          readOnly: true
        title:
          type: string
          maxLength: 100
        slug:
          type: string
          maxLength: 140
          pattern: ^[-a-zA-Z0-9_]+$
        description:
          type: string
        short_description:
          type: string
          maxLength: 200
        category:
          type: string
          format: uuid
        category_detail:
          allOf:
          - $ref: '#/components/schemas/AdminProjectCategory'
          readOnly: true
        location:
          type: string
          maxLength: 120
        location_governorate:
          type: string
          maxLength: 120
        goal_amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,2})?$
        funded_amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,2})?$
          readOnly: true
        funding_account:
          type: string
          readOnly: true
        funding_reached_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        pending_payment_deadline:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        funding_finalized_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        funding_finalized_by:
          type: string
          format: uuid
          readOnly: true
          nullable: true
        minimum_investment:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
        expected_roi:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,2})?$
        cost_items: {}
        faqs: {}
        funding_period_days:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
        start_date:
          type: string
          format: date-time
        end_date:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
          readOnly: true
        is_verified:
          type: boolean
        verified_by:
          type: string
          format: uuid
          nullable: true
        verified_by_detail:
          allOf:
          - $ref: '#/components/schemas/AdminRelatedUser'
          readOnly: true
        verified_at:
          type: string
          format: date-time
          nullable: true
        verification_notes:
          type: string
        business_plan:
          type: string
          format: uri
          nullable: true
        financial_projections:
          type: string
          format: uri
          nullable: true
        ownership_proof:
          type: string
          format: uri
          nullable: true
        cover_image:
          type: string
          format: uri
          nullable: true
        images:
          type: array
          items:
            $ref: '#/components/schemas/AdminProjectImage'
          readOnly: true
        video_url:
          type: string
          format: uri
          maxLength: 200
        ai_classified_category:
          type: string
          maxLength: 100
        ai_confidence_score:
          type: string
          format: decimal
          pattern: ^-?\d{0,1}(?:\.\d{0,2})?$
          nullable: true
        ai_classification_at:
          type: string
          format: date-time
          nullable: true
        ai_generated_summary:
          type: string
        milestone_count:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
        repayment_status:
          $ref: '#/components/schemas/RepaymentStatusEnum'
        total_repaid:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,2})?$
        next_repayment_date:
          type: string
          format: date
          nullable: true
        view_count:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
        investor_count:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
        rating:
          type: string
          format: decimal
          pattern: ^-?\d{0,1}(?:\.\d{0,2})?$
        reviews_count:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
        supporting_documents:
          type: array
          items:
            $ref: '#/components/schemas/AdminProjectDocument'
          readOnly: true
        milestones: {}
        days_left:
          type: string
          readOnly: true
        funding_percent:
          type: string
          readOnly: true
        pending_edit_request:
          type: string
          readOnly: true
        implementation_complete:
          type: string
          readOnly: true
        updates:
          type: string
          readOnly: true
        deleted_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - category
      - category_detail
      - created_at
      - days_left
      - description
      - entrepreneur
      - entrepreneur_detail
      - funded_amount
      - funding_account
      - funding_finalized_at
      - funding_finalized_by
      - funding_percent
      - funding_reached_at
      - goal_amount
      - id
      - images
      - implementation_complete
      - location
      - pending_edit_request
      - pending_payment_deadline
      - short_description
      - status
      - supporting_documents
      - title
      - updated_at
      - updates
      - verified_by_detail
    AdminProjectCategory:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
          maxLength: 80
        slug:
          type: string
          maxLength: 100
          pattern: ^[-a-zA-Z0-9_]+$
        description:
          type: string
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - created_at
      - id
      - name
      - updated_at
    AdminProjectDocument:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        project:
          type: string
          format: uuid
        file:
          type: string
          format: uri
        title:
          type: string
          maxLength: 120
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - created_at
      - file
      - id
      - project
      - title
      - updated_at
    AdminProjectImage:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        project:
          type: string
          format: uuid
        image:
          type: string
          format: uri
        alt_text:
          type: string
          maxLength: 160
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - created_at
      - id
      - image
      - project
      - updated_at
    AdminRelatedUser:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        email:
          type: string
          format: email
          maxLength: 254
        full_name:
          type: string
          maxLength: 150
        user_type:
          $ref: '#/components/schemas/UserTypeEnum'
        business_name:
          type: string
          maxLength: 150
      required:
      - email
      - full_name
      - id
...
    RepaymentStatusEnum:
      enum:
      - on_track
      - delayed
      - completed
      type: string
      description: |-
        * `on_track` - On track
        * `delayed` - Delayed
        * `completed` - Completed
...
    UserTypeEnum:
      enum:
      - investor
      - entrepreneur
      - admin
      type: string
      description: |-
        * `investor` - Investor
        * `entrepreneur` - Entrepreneur
        * `admin` - Admin
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT